refactor controller tests
mocking activegate reconciler and adding a test

rename for clarity

remove redundant code

rename for clarity

moves test to unit-test of proxy/reconciler_test.go

moves tests to respective component

fix grammar in comments

implemens PR feedback
