+++ .mockery.yaml
@@ -37,6 +37,9 @@ packages:
   github.com/Dynatrace/dynatrace-operator/pkg/controllers/dynakube/connectioninfo:
     interfaces:
       Reconciler:
+  github.com/Dynatrace/dynatrace-operator/pkg/controllers/dynakube/deploymentmetadata:
+    interfaces:
+      Reconciler:
   github.com/Dynatrace/dynatrace-operator/pkg/injection/codemodule/installer:
     interfaces:
       Installer:
