+++ test/mocks/pkg/controllers/dynakube/istio/reconciler.go
@@ -6,7 +6,6 @@ import (
 	context "context"
 
 	dynakube "github.com/Dynatrace/dynatrace-operator/pkg/api/v1beta1/dynakube"
-
 	mock "github.com/stretchr/testify/mock"
 )
